Correctly handle duplicate WellKnownType values
The WellKnownType enumeration contains several named items that have the
same underlying value.  Both First and Last share a value with an actual
type entry in the enumeration.  A ToString call on those values can
return either the type name or First / Last, the runtime does not define
which will happen.

This comes into play in the static constructor of WellKnownTypes where
we assert that the enumeration and s_metadataNames table stay in sync.
The code already accounts for First being ambiguous but failed to for
Last.

This change removes the assumption for Last as well.  It also expands
out the assert as it was getting to difficult to follow when I added
another level of ternary expressions into the mix.