fix a bug where we gave wrong value for length in tagger
this is reported by editor team where we returning wrong tags for given span. it looks like we gave end position as length. this is a second time we did this mistake. probably because this code was based on the code that had this issue originally.

to prevent this from happening again, I added assert in base tagger class.

* this is hard to do in unit test since each unit test tests one specific tagger. if we forgot to add unit test to tests this specific issue for new tagger added, it will go unnoticed again.