Handle failing conversion in synthesized code
We used to assume that the conversion from the static return type of an
intrinsic method to the runtime type of the corresponding pseudo-variable
could never fail.  However, in the presence of not-yet-loaded modules (esp
corlib), it can.  Propagate the resulting diagnostic out to a point where
it can be used to attempt loading of the not-yet-loaded module.

Caveat: This doesn't work as well in VB, which tends to encounter
MissingCorLibrarySymbol before it encounters a missing assembly with a
useful identity (GitHub #597).

Fixes DevDiv #1124725.