Handle TaskCancelledException in AnalyzerManager.GetAnalyzerActionsAsync
AnalyzerManager maintains a cache of per-analyzer tasks to compute sessions and compilation analyis scopes. If multiple clients simultaneously ask for the analyzer actions for a given analyzer, then all of them receive the same task. If the thread that created this task is now cancelled, all other clients get a TaskCancelledException with a different cancellation token then theirs and the exception goes unhandled, bringing down VS.

Fix is to catch this task cancelled exception in AnalyzerManager, remove the entry for the the cancelled task in the scope map and then re-attempt computing scope, if our cancellelation has not fired.