Re-compile EE binaries against Phone references
Even though they share the same APIs, the desktop and phone versions of
the Concord binaries have different strong names.  In order to support
both scenarios, we have to produce each of our EE (product) binaries
twice: once referencing the desktop assemblies and again referencing the
phone assemblies.

Catch 1: There are no reference assemblies for the phone versions of the
Concord binaries, so we explicitly re-assemble the same reference IL with
a different strong name key.

Catch 2: We can't put two copies of each assembly in a single output
directory, so we reuse the existing Phone subdirectory and put both the
new references and the new outputs in there.  We do not run our tests
against these assemblies since they differ only in strong name
(references).