Handle AssemblyResolve events even when the requesting assembly is unknown
Since we handle loading analyzer assemblies ourselves (to avoid locking
the file on disk and the need to shadow copy them) we must also hook the
AssemblyResolve event and handle finding and loading their dependencies.
Currently we only handle the event if the event arguments specify the
requesting assembly (via `ResolveEventArgs.RequestingAssembly`) and the
requesting assembly is known to us, i.e. we were responsible for loading
it.

Unfortunately `ResolveEventArgs.RequestingAssembly` is more of a hint
than anything else, and there are circumstances where it can be null. As
such we sometimes fail to load a dependency needed by an analyzer, and
it doesn't run.

The fix here is to attempt to load the assembly even when
`RequestingAssembly` is null. Instead of looking next to the requesting
assembly, we will look next to every assembly we have already loaded for
one with the correct identity, and load and return that.

Fixes #635.