Ensure AsyncLazy throws correct OperationCanceledException
When you call GetValue or GetValueAsync, it's possible that the
method may throw OperationCanceledException if you are the only
requestor and the request is cancelled during your call. In this
case we need to ensure that the token associated with the exception
is the CancellationToken that you passed in instead of the computation
token that we created.