Fix to VB pretty-listing of line continuation characters around string interpolations
The VB pretty-lister shouldn't remove line continuation characters after
the opening brace or before the closing brace of an interpolated string.
In addition, I found that the pretty-lister was greedily removing line
continuation characters in error conditions where a missing token was
inserted by the compiler to represent the list token of a statement.