Show 'else' within Ifs within Elses
When recommending "Else" we walk all the way up the tree to determine
if the position is nested inside an Else block. This means we don't
recommend "Else" inside an If block nested within an Else block. When
walking up the tree, we should stop at the If/ElseIf block.