Combine two events that really represent the same thing
AnalyzerItemTracker has an event that fires when the select hierarchy
changes, and an event that fires when the selected item ID changes.
These really represent two aspects of the same underlying change: the
selected item in the Solution Explorer has changed. So combine these
into a single event that fires when either the hierarchy or item ID
changes.