Use TriggerDiagnosticDescriptor in RenameTracking
Fixes #466 "Use a 'TriggerDiagnosticDescriptor' for the rename tracking
diagnostic"

Updates the RenameTrackingDiagnosticAnalyzer to use a
TriggerDiagnosticDescriptor, which prevents the diagnostic title from
appearing in the preview area of the lightbulb. We now pass the
RenameFrom and RenameTo names as part of the Diagnostic's property bag
and construct the localized codefix title based on these in the codefix
itself.