Backspace completion should map the caret to the subject buffer
Completion on backspace uses the caret position to determine if it should trigger, which involves inspecting the subject buffer. It did not map the text view caret position onto the subject buffer position. This resulted in crashes in projection scenarios when the surface buffer was longer than the subject buffer.