Update context menu item state lazily
Right now we update a bunch of state within our Solution Explorer
context menus on every change to the set of selected items in the
Solution Explorer. This state includes things like which items are
visible, which are enabled, and which are checked. In most cases it
would be better to simply wait until a context menu is about to be
displayed, and then go update the item state.

The big change here is that ContextMenuController now updates the menu
item state before showing the menu; this requires that the
AnalyzersCommandHandler create the ContextMenuController instances since
it owns the menu items.

Most of this change is simply about piping the ContextMenuControllers
from the AnalyzersCommandHandler down to the individual
AnalyzersFolderItems/AnalyzerItems/DiagnosticsItems. I also had to add
an IAnalyzerCommandHandler interface so that I could create a no-op
implementation for the unit tests.