Move our extra solution to work with CTP6
This change moves Roslyn2013.sln to work with Visual Studio 2015 CTP6.
This allows customers, and Jenkins, to build and test significantly more
code than was available with the old solution.

This is still not as full Roslyn.sln because not everything is buildable
against CTP6.  In particular our Visual Studio components and all of
their dependencies can't build due to API changes we depend on that came
into being after CTP6.  It is much closer though than our previous
solution.

As a part of this change I also renamed Roslyn2013.sln to
RoslynLight.sln.  The 2013 moniker was no longer valid and lacking a
better word I chose light (it has less projects and hence is lighter).