Unskip tests for issue #204
When a request to AsyncLazy gets a cancellation, we asynchronously
cancel the underlying Task, since any continuations might run inline
and cause surprises. This meant that if cancellation happened during
a call to GetValueAsync or GetValue where the underlying computation
was synchronously completed, there was a race where the task that
gets returned might or might not be cancelled. It didn't strictly
matter -- since cancellation is best-effort it was fine by the contract.
But to test an earlier bug, we forced it into the non-cancelling
case by simply stopping up the thread pool so the asynchronous
cancellation never happened. That approach was really evil, so we try
something else.

Now, we adjust the tests to assert the slightly stronger statement that
once a computation runs and returns a value, it's cached, even if
the value is not handed out. (The implementation already did this,
but we now assert it.) We also tweak the cancellation handling to
cancel the returned task more aggressively. This avoids the race
between async cancellation and synchronous value returning in the
context of the unit test, so we can assert behavior better. In general
there's no guarantee, but it's now guaranteed "enough" our test can
test the other behaviors well.