Show rule set severity in the context menu
When you right-click on a diagnostic in the Solution Explorer and open
the "Set Rule Set Severity" sub menu, the check marks actually show you
the effective severity of the rule. This takes into account not just
what is in the rule set file, but also the default severity and any
relevant project options.

This is misleading for a couple of reasons. For one thing, we may show a
checkmark on a severity even though the rule doesn't appear in the rule
set file at all. Worse, choosing a different severity may update the
rule set but leave the checkmark where it is (because some other setting
overrides the rule set) and quite likely confusing the user.

This change updates the checkmarks to truly show what is in the rule set
file, and only what is in the rule set file.