Run csc.exe from the MSBuild directory
The AnalyzerDependencyCheckerTests currently build a bunch of assemblies
for testing purposes. It assumes it will find csc.exe next to the
assembly contain the unit tests. This is true on systems that have built
Roslyn from source (local dev boxes and Jenkins check-in systems, for
example) but is not true for "standalone" test runs that just use
binaries built elsewhere.

Since these tests compile very simple assemblies and do not depend on
any bleeding-edge language features, we can simply uses the csc.exe from
the MSBuild directory; that one should always be available.