Prototype: Ask for bytes of not-yet-loaded winmd
When we try to bind something that is from a not-yet-loaded winmd, we
generally see ErrorCode.ERR_DottedTypeNameNotFoundInNS because the
namespace Windows.X is only provided by Windows.X.winmd.  When this
occurs, heuristically ask the debugger for an assembly with name Windows.X
and content type AssemblyContentType.WindowsRuntime.

Caveat: I tried this with Windows.Storage.winmd, which exists on my box,
but the debugger didn't return bytes.  Either this approach isn't sound,
or debugger work is required.