C# command line options should override rule sets
If a diagnostic is disabled by default, you have two ways of turning it
on in the C# compiler:

1. Add a command line switch, e.g., `/WarningAsError+:MyDiagnosticID`.
2. Add a rule to the .ruleset file, e.g., `<Rule Id="MyDiagnosticID"
Action="Warn" />`.

The former turns the diagnostic on and always makes it an error. The
latter turns it on and sets its severity to whatever the `Action`
attribute specifies.

We run into problems when you use a rule set to turn on MyDiagnosticID
and make it a warning, and then specify `/WarnAsError+` to turn all
warnings into errors. Currently, MyDiagnosticID remains a warning
because since the rule set supplies a specific option ("MyDiagnosticID
is a warning") and specific options always override general options
("All warnings should be treated as errors").

What we want instead is for command line options to override the options
in rule sets. With this change, when you specify `/WarnAsError+` on the
command line we'll now go through the options from the rule set file and
turn any warnings into errors. A subsequent `/WarnAsError-:MyDiagnostic`
will restore the severity to whatever was in the rule set (or Default,
if it isn't in the ruleset).