VB command line options should override rule sets
VB has the same problem where a rule turned on in a rule set file will
not be promoted to errors as you expect when using `/WarnAsError`. To
fix that, `/WarnAsError` now explicitly promotes warnings from rule set
files into errors. To accomplish this without breaking existing behavior
around `/WarnAsError` and `/NoWarn`, we implement the following
hierarchy of settings:

1. Specific `/NoWarn` arguments, in order
2. Specific `/WarnAsError` arguments, in order
3. General `/NoWarn` and `/WarnAsError` arguments, in order
4. Rule set file settings