reduce chance to report blank entries in error list.
current design of error list is that when solution is changed, we will momentally get out of sync but eventually catch up and make error list and solution in sync. during those out-of-sync period, it is possible for error list to show blank entries in the list. this change is to reduce those cases.

more correct way to do this is making error list and diagnostic service to share same version. that requires a bit of change in both side, so I am trying cheap way first. since at the end, two should get into sync eventually even without this change, cheap way probably is enough.