Debugger IntellIiSense and linked files
Since we did the shared project work, completion computes results for
every related document. Since debugger intellisense only replaced the
Text of the context document and not its related documents, completion
in a linked file in debugger intellisense could easily query invalid
caret positions and crash with ArgumentOutOfRange.

To fix this, we replace the text for all the related documents of the
context document. I believe this is consistent with normal handling of
linked files where we ensure that linked files always have the same
TextContainer.