Html should be the default output for tests run
When running tests from the command line the default output for xUnit
should be HTML, not XML.  The XML output only exists to facilitate
better display in Jenkins and is controlled by the -xml command line
option.  When running locally developers should see the HTML output as
it is much easier to read.

The current behavior is just a bug I added some time ago.