Fix ContainedLanguage support for Workflow v1 XOML files
Workflow XOML files call into IVsContainedLanguageFactory.GetLanguage(),
passing an IVsTextBufferCoordinator. However, that
IVsTextBufferCoordinator is constructed differently than the one we get
from Venus and breaks some assumptions in the C#/VB language services.

The critical assumption that XOML breaks is that the primary buffer is
an IProjectionBuffer. This is not true for XOML. Instead, it is simply a
non-projected buffer with the HTML content type.  So, our code attempts
to cast to an IProjectionBuffer and we crash before ever creating a
ContainedDocument and adding it to the workspace. To fix this, we need
to break the IProjectionBuffer assumption and update any calling code
depending on projections to dynamically cast to IProjectionBuffer first.
This unblocks several Workflow v1 scenarios.

Next, it turns out that XOML never calls SetHost() on the
IVsContainedLanguage that we return. So, there's never an
IVsContainedLanguageHost hooked up and calling code needs to be updated
to guard against null hosts.