Only show check mark if the severities are the same
If the user selects a bunch of diagnostics with differing severities in
Solution Explorer, right-clicks,  and opens the "Set Rule Set Severity"
submenu they will see a check mark next to every applicable severity.
E.g., if one rule is an error and one is a warning, then both Error and
Warn will be checked.

The general feeling is that this will be confusing to users. Instead,
only show a check mark when all the selected rules share the same
severity.