Getting Parent of SyntaxNodeOrToken does not need conversion to Node or Token.
Fix internal TFS bug 1130555 : Getting Parent of SyntaxNodeOrToken
does not need conversion to Node or Token.

When converting a SyntaxNodeOrToken to a Node or Token use IsNode or IsToken.

In this particular instance, we have introduced the check in
InlIneTemporaryCodeRefactoringProvider to prevent crash. We will still
have to find the scenario that triggered this situation.