Add the CoreCompile targets for C# & VB to MSBuildTask project
This extracts the CoreCompile target from the Microsoft.CSharp.CurrentVersion
and Microsoft.VisualBasic.CurrentVersion targets files and creates two new
targets files, Microsoft.CSharp.Core and Microsoft.VisualBasic.Core, to contain
it.

This is necessary because the CoreCompile target contains the actual invocation
of the Csc/Vbc tasks, which the MSBuildTask assembly contains. Any changes in
the public surface area of the tasks need to be reflected in the targets. These
targets should eventually be inserted into the MSBuild install with the task as
the compiler public entry point.