Match native compiler PDB behavior
This causes the compiler to match the native one in terms of PDB
behavior.  It will only pass along the Stream value to Initialize in the
cases where a Stream is explicitly provided.  In the other cases,
basically the raw command line, it will pass along only the file name.
This puts the responsibility of creating / writing the file solely in
ISymUnmanagedWriter as it previously was via WritePdbToOutput