Unify Stream passing API
This unifies the Stream passing API to prefer EmitStreamProvider over
Stream in the Emit APIs.  Even though we are not currently taking
advantage of the delay creation mechanism for the command line compiler
there was a desire to have consistent parameter values where possible.
The advantage is it lets us switch to making XML documentation delay
created as well with a very small change in the future.