Ensure a file name for PDB
The ISymUnmanagedWriter2 interface does not require a file name be
passed.  It can successful emit a PDB without it.  The only method which
actually requires it is GetDebugInfo.

The Compilation type will call this member frequently during PDB
creation.  Hence if we have a Stream only we need to manufacture a name
to use for it.