Ensure AnalyzerDriver completes during Compilation
Today the AnalyzerDriver is only guaranteed to be completed after a
Compilation if the compilation made it to the Emit phase.  If the
compiler exited early, say to method body binding failures, the compiler
would not produce all of the compilation events and hence the analyzer
driver would be left forever waiting for them to complete.