Introduce a new AbstractEncProject for CSharpProject and VisualBasicProject to extend.
This project does the appropriate Enc hookup, and implements the interfaces that the debugger
QIs for to do Enc interaction.

For C# and VB, nothing will change.  For TypeScript, we will continue subclassing AbstractProject.
By doing this, we will not interact with the Enc system at all.  This will help avoid undesirable
behavior (like having our files become read-only when entering debugging).