Added suppression attributes for FxCop CA2001:AvoidCallingProblematicMethods warnings
FxCop flagged our use of GC.Collect and System.Runtime.InteropServices.DangerousGetHandle.
The cases seems to be justified uses.

We use GC.Collect when we expect long periods of inactivity. This is the rare case where we might know better than GC that it is a good time to trim garbage.

We use System.Runtime.InteropServices.DangerousGetHandle for logging purposes. Nothing is dependent on the actual value returned.