Resolve rename conflicts in nameof in attributes
The expander can operate on a specific set of known nodes, so when
rename finds a conflict in a particular node, it uses the span of a
parent that the expander knows how to handle. For normal
ArgumentSyntaxes, this is covered by being within a StatementSyntax.
However, nameof within an attribute constructor isn't in a statement, so
rename needs to expand to the span of the attribute.