Handle not-yet-loaded projects in GetENCBuildState
VsENCRebuildableProjectImpl.GetNECBuildState was not accounting for the
possibility that there might not (yet) be a roslyn project for a given VS
project.  If we ever encounter this, we should just report that the
project is unchanged.

We have a number of dumps for this issue, but no concrete repro steps.