Qualify member accesses with type name in nameof
Fixes #1193

During expansion, directly accessed members need to be qualified with
"this." or "TypeName.". Prior to this change, we always qualified
instance members with "this." and static members with "TypeName.", but
this technique fails in nameof contexts where an instance member can be
referenced from a static method, resulting in an improper "this."
qualification. We can only legally use "this." when both the member
being accessed and the context from which we are accessing it are
non-static. However, this change updates the expanders to always use
"TypeName." qualification in nameof contexts regardless of the
staticness of either member because it is always legal and feels more
natural in nameof expressions (where no actual value is being accessed).