Convert FakeSign to a NuGet reference
Lately our developer builds have gotten a bit flaky due to the
dependency on FakeSign.  This is both a tool we create and consume
during build.  In order to do this properly every tool which depends on
FakeSign should list it as a dependency.  This ensures it will be built
before the project which consumes it.

Unfortunately this is not the case.  Virtually nothing lists it as a
dependency and hence we are often broken during build due to ordering
issues.  This fixes the problem by hosting FakeSign as a NuGet package.

Note: the long term plan is to have the compiler support oss signing
directly from the command line.  At that point the majority of our
projects will employ fake / oss signing via a command line switch.  The
FakeSign tool will still be necessary for IL projects like Concord
though hence this is not wasted work.