Make the IWorkCoordinatorPriorityService async and appropriately cascade out the asynchrony.
Originally this was synchronous because it was used through codepaths that all were
inherently synchronous.  Over the last year the WorkCoordinator has become increasingly
async, and we're now at the point where this API can be async too.  This is a win for
TypeScript because we have to currently implement this API in blocking fashion.  This
has shown up in profiles, and making it async should be a nice win for us.